fix(cloister): break stall sweeper service cycle
Issue: PAN-3501
